% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_datasets.R
\docType{data}
\name{bodyfat}
\alias{bodyfat}
\title{Body Fat}
\format{
A data frame with use 252 observations (rows) and 15 variables (columns). Each row is the measure for a man.
\describe{
\item{density}{A numeric column with the density determined from underwater weighing.}
\item{bodyfat}{A numeric column with the percent body fat from Siri's (1956) equation (Percent: 0 - 100).}
\item{age}{A numeric column with age in years.}
\item{weight}{A numeric column with weight in pounds (lbs).}
\item{height}{A numeric column with height in inches.}
\item{neck}{A numeric column with the neck circumferences measurements in cm.}
\item{chest}{A numeric column with the chest circumferences measurements in cm.}
\item{abdomen}{A numeric column with the abdomen circumferences measurements in cm.}
\item{hip}{A numeric column with the hip circumferences measurements in cm.}
\item{thigh}{A numeric column with the thigh circumferences measurements in cm.}
\item{knee}{A numeric column with the knee circumferences measurements in cm.}
\item{ankle}{A numeric column with the ankle circumferences measurements in cm.}
\item{bicep}{A numeric column with the bicep circumferences measurements in cm.}
\item{forearm}{A numeric column with the forearm circumferences measurements in cm.}
\item{wrist}{A numeric column with the wrist circumferences measurements in cm.}
}
}
\source{
The data was collected by A. Garth Fisher from the study of Penrose et al (1985)
This data and column descriptions were pulled from StatLib (\url{http://lib.stat.cmu.edu/datasets/bodyfat}).

Penrose, K. W., Nelson, A. G., & Fisher, A. G. (1985). Generalized body composition prediction equation for men using simple measurement techniques. Medicine & Science in Sports & Exercise, 17(2), 189.
}
\usage{
bodyfat
}
\description{
Predict percentage of body fat determined by underwater weighing (a more accurate measure of body fat) using various body circumference measurements (which are more convenient measures than underwater weighing) for 218 men.
}
\keyword{datasets}
