% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_datasets.R
\docType{data}
\name{cities_orig}
\alias{cities_orig}
\title{German City Populations}
\format{
A data frame with 83 observations (rows) and 11 variables (columns):
\describe{
\item{city}{A character column with the name of the city}
\item{pop}{A numeric column with the 1993 city population}
\item{soccer}{A numeric column with a boolean flag for whether the city has a team in the major soccer league}
\item{state_capital}{A numeric column with a boolean flag for whether the city is a state capital}
\item{east_germany}{A numeric column with a boolean flag for whether the city was formerly in East Germany}
\item{industrial_belt}{A numeric column with a boolean flag for whether the city was in the industrial belt}
\item{license_plate}{A numeric column with a boolean flag for whether the abbreviation for the city on license plates is one-letter long.}
\item{train}{A numeric column with a boolean flag for whether the city is on the Intercity train line}
\item{exposition_site}{A numeric column with a boolean flag for whether the city once an exposition site}
\item{national_capital}{A numeric column with a boolean flag for whether the city is the national capital of Germany}
\item{university}{A numeric column with a boolean flag for whether the city home to a university}
}
}
\source{
Fischer Welt Almanach (Fischer World Almanac). (1993). Frankfurt, Germany: Fischer.
This data can be found in the appendix of:
Gigerenzer, G., & Goldstein, D. G. (1996). Reasoning the fast and frugal way: models of bounded rationality. Psychological review, 103(4), 650.
Found here: \url{http://www.dangoldstein.com/papers/FastFrugalPsychReview.pdf}
}
\usage{
cities_orig
}
\description{
Predict populations of the 83 German cities with at least 100,000 inhabitants based on whether each city has a soccer team, university, intercity train line. exposition site, etc.
}
\note{
This data contains two data transcription errors--matching the dataset used in the 1996 Gigerenzer, Goldstein paper.
\itemize{
\item The state capital (\code{state_capital})entry for the city of Dresden
\item The soccer team (\code{soccer}) entry for the city of Krefeld
}
}
\keyword{datasets}
