% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation_datasets.R
\docType{data}
\name{fuel}
\alias{fuel}
\title{Fuel Consumption}
\format{
A data frame with 48 observations (rows) and 9 variables (columns). Each row is a state.
\describe{
\item{state}{A character column with the 2 letter state abbreviation}
\item{pop}{A integer column with the 1971 population in thousands}
\item{tax}{A numeric column with the 1972 motor fuel tax rate in cents/gallon}
\item{lic_n}{A integer column with the 1971 number of licensed drivers in thousands}
\item{income}{A numeric column with the 1972 per capita income in thousands of dollars}
\item{road}{A numeric column with the 1971 miles of federal-aid primary highways in thousands}
\item{fuel_n}{A integer column with the 1972 fuel consumption in millions of gallons}
\item{lic_pc}{A numeric column with the percent of the population with a driver's license, calculated from \code{100*(lic_n/pop)}}
\item{fuel_pc}{A integer column with the fuel consumption in gallons per capita, calculated from \code{1000*(fuel_n/pop)}}
}
}
\source{
The data were collected by Christopher Bingham for the American Almanac for 1974, except for fuel consumption, which was given in the 1974 World Almanac. The dataset was reported in Weisberg (1985).
This data was pulled from StatLib (\url{http://lib.stat.cmu.edu/datasets/alr}) under the name \code{alr35}.
Column descriptions from Weisberg (1985).

Weisberg, Sanford. "Simple linear regression." Applied Linear Regression (1985)
}
\usage{
fuel
}
\description{
ADD FUEL CONSUMPTION DESCRIPTION
}
\keyword{datasets}
