% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economics_datasets.R
\docType{data}
\name{housing}
\alias{housing}
\title{Housing Prices in Erie, Pennsylvania}
\format{
A data frame with 27 observations (rows) and 10 variables (columns). Each row is a house that was sold in Erie, Pennsylvania.
\describe{
\item{taxes}{A numeric column with the current taxes (i.e. local, school, and county) in dollars / 100}
\item{bathrooms}{A numeric column with the number of bathrooms}
\item{lot_sq}{A numeric column with the lot size in square feet / 1000}
\item{living_sq}{A numeric column with the living space in square feet / 1000}
\item{garage}{A numeric column with the number of garage spaces}
\item{rooms}{An integer column with the number of rooms in the house}
\item{bedrooms}{An integer column with the number of bedrooms in the house}
\item{age}{An integer column with the age of the house in years}
\item{fireplaces}{An integer column with the number of fireplaces in the house}
\item{price}{A numeric column with the actual sale price in dollars / 1000}
}
}
\source{
Dataset was published in Narula, Wellington (1977) and reported in Weisberg (1985)
This data was pulled from StatLib (\url{http://lib.stat.cmu.edu/datasets/alr}) under the name \code{alr241}.
Column descriptions from Weisberg (1985).

Narula, S. C., & Wellington, J. F. (1977). Prediction, linear regression and the minimum sum of relative errors. Technometrics, 19(2), 185-190.
Weisberg, Sanford. "Simple linear regression." Applied Linear Regression (1985)
}
\usage{
housing
}
\description{
The data from 27 houses sold in Erie, Pennsylvania that were sold.
}
\keyword{datasets}
