% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economics_datasets.R
\docType{data}
\name{landrent}
\alias{landrent}
\title{Land Rent}
\format{
A data frame with 67 observations (rows) and 5 variables (columns). Each row is a county in Minnesota. Only counties with "appreciable rented farmland are included."
\describe{
\item{alf_rent}{A numeric column with the average rent per acre planted to alfalfa}
\item{till_rent}{A numeric column with the average rent paid for all tillable land}
\item{cows}{A numeric column with the density of dairy cows (number per square mile)}
\item{pasture}{A numeric column with the proportion of farmland used as pasture}
\item{alf_grow}{A numeric column with a boolean flag if liming is required to grow alfalfa}
}
}
\source{
The data was provided by Douglas Tiffany and reported in Weisberg (1985).
This data was pulled from StatLib (\url{http://lib.stat.cmu.edu/datasets/alr}) under the name \code{alr162}.
Column descriptions from Weisberg (1985).

Weisberg, Sanford. (1985). Applied linear regression. New York: Wiley.
}
\usage{
landrent
}
\description{
"Alfalfa is a high protein crop that is suitable feed for dairy cows. It is thought that rent for land planted to alfalfa relative to rent for other agricultural purposes would be higher in areas with a high density of dairy cows and rent would be lower in counties where limin in required, since that would mean additional expense."
Weisberg (1985)
}
\keyword{datasets}
