% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biology_datasets.R
\docType{data}
\name{manure}
\alias{manure}
\title{Cow manure}
\format{
A data frame with 20 observations (rows) and 8 variables (columns). Each row is a day when the sample was measured.
\describe{
\item{day}{A numeric column with the day when the sample was measured}
\item{bod}{A numeric column with the biological oxygen demand}
\item{tkn}{A numeric column with the total Kjeldahl nitrogen}
\item{ts}{A numeric column with the total solids}
\item{tvs}{A numeric column with the total volatile solids}
\item{cod}{A numeric column with the chemical oxygen demand}
\item{o2up}{A numeric column with the oxygen uptake}
\item{log_o2up}{A numeric column with the log transformation of the \code{o2up} variable}
}
}
\source{
Dataset was published in Moore (1975) and reported in Weisberg (1985)
This data was pulled from StatLib (\url{http://lib.stat.cmu.edu/datasets/alr}) under the name \code{alr222}.
Column descriptions from Weisberg (1985).

Moore, J. A. (1975). Total biochemical oxygen demand of animal manures. University of Minnesota.
Weisberg, Sanford. (1985). Applied linear regression. New York: Wiley.
}
\usage{
manure
}
\description{
description
}
\keyword{datasets}
