% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_datasets.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Mortality in US Cities}
\format{
A data frame with 60 observations (rows) and 16 variables (columns):
\describe{
\item{prec}{An integer column with average annual precipitation (in inches)}
\item{jan_temp}{An integer column with the average January temperature (F)}
\item{jul_temp}{An integer column with the average July temperature (F)}
\item{over_65}{A numeric column with the percent of 1960 SMSA population aged 65 or older (Percent: 0 - 100)}
\item{household}{A numeric column with the average household size}
\item{educ}{A numeric column with the median school yers completed by those over 22}
\item{hous}{A numeric column with the percent of housing unites which are sound and with all facilities (Percent: 0 -100)}
\item{dens}{An integer column with population per sq. mile in urbanized areas, 1960}
\item{non_white}{A numeric column with the percent non-white population in urbanized areas, 1960 (Percent: 0 -100)}
\item{white_collar}{A numeric column with the percent occupaitions that are white collar (Percent: 0 - 100)}
\item{poor}{A numeric column with the percent of households with less than $3000 in annual income (Percent: 0 -100)}
\item{hc}{An integer column with the relative hydrocarbon pollution potential}
\item{nox}{An integer column with the relative nitric oxides potential}
\item{so}{An integer column with the relative sulphur dioxide}
\item{humid}{An integer column with the annual average percent relatve humidity at 1 pm (Percent: 0 -100)}
\item{mort}{A numeric column with the total age-adjusted mortality rate per 100,000}
}
}
\source{
McDonald, G. C, & Schwing, R. C. (1973). Instabilities of regression estimates relating air pollution to mortality. Technometrics, 15, 463—482;
This data and its summary is pulled from StatLib: \url{http://lib.stat.cmu.edu/datasets/pollution}
}
\usage{
mortality
}
\description{
Predict the mortality rate in 20 U.S. cities given the average January temperature, pollution level, and percentage of nonwhites, etc.
}
\keyword{datasets}
