% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_datasets.R
\docType{data}
\name{obesity}
\alias{obesity}
\title{Obesity at age 18}
\format{
A data frame with 136 observations (rows) and 12 variables (columns). Each row is a study participant.
\describe{
\item{sex}{An integer column with a boolean flag if the participant is male or female. Coded 0 for male, 1 for female.}
\item{weight2}{A numeric column with the participant's weight at age 2 in kg.}
\item{height2}{A numeric column with the participant's height at age 2 in inches.}
\item{weight9}{A numeric column with the participant's weight at age 9 in kg.}
\item{height9}{A numeric column with the participant's height at age 9 in inches.}
\item{leg9}{A numeric column with the participant's leg circumference at age 9 in cm.}
\item{strength9}{A numeric column with a composity measure of strength at age 9. Higher values = stronger.}
\item{weight18}{A numeric column with the participant's weight at age 18 in kg.}
\item{height18}{A numeric column with the participant's height at age 18 in inches.}
\item{leg18}{A numeric column with the participant's leg circumference at age 18 in cm.}
\item{strength18}{A numeric column with a composity measure of strength at age 18. Higher values = stronger.}
\item{soma}{A numeric column with the seven-point scale of the participant's body type at age 18. Higher values = larger.}
}
}
\source{
Dataset was published on he longitudinal monitoring of the Berkeley Guidance Study, Tuddenham & Snyder (1954), A data excerpt was reported in Weisberg (1985).
This dataset was pulled from the \code{alr3} package under the name \code{BGSall}.
Column descriptions from Weisberg (1985).

Tuddenham, R. D. (1954). Physical growth of California boys and girls from birth to eighteen years. University of California publications in child development, 1, 183-364.
Weisberg, Sanford. (1985). Applied linear regression. New York: Wiley.
}
\usage{
obesity
}
\description{
Predict fatness at age 18 children based on body measurements from age 2 to age 18. The body measurements included height, weight, leg circumference, and strength.
}
\note{
The Czerlinski et. al. (1999) used a subset (46 participants) of this dataset--found in the Weisberg (1985) excerpt.
}
\keyword{datasets}
