% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economics_datasets.R
\docType{data}
\name{salaries}
\alias{salaries}
\title{Professor Salaries}
\format{
A data frame with 52 observations (rows) and 6 variables (columns). Each row is a professor.
\describe{
\item{sex}{A integer column with the sex of the professor.}
\item{rank}{A integer column with the rank of the professor; Coded 1 for Assistant Professor, 2 for Associate Professor, and 3 for Full Professor.}
\item{year_r}{A integer column with the number of years in the professor's current rank.}
\item{degree}{A integer column with the highest degree of the professor. Coded 1 for Doctorate and 2 for Masters.}
\item{year_d}{A integer column with the number of years since the highest degree was earned.}
\item{salary}{A integer column with the professor's salary in the current academic year in dollars.}
}
}
\source{
The data was reported in Weisberg (1985).
This data was pulled from StatLib (\url{http://lib.stat.cmu.edu/datasets/alr}) under the name \code{alr194}.
Column descriptions from Weisberg (1985).

Weisberg, Sanford. (1985). Applied linear regression. New York: Wiley.
}
\usage{
salaries
}
\description{
This dataset contains the "salary and other characteristics of all faculty in a small Midewestern college. The data have been collected for presentation in legal proceedings for which discrimination against women in salary was at issue. All persons in the data hold tenured or tenure track positions; temporary faculty are not included. The data were collected from personal files..."
Weisberg (1985)
}
\keyword{datasets}
