% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biology_datasets.R
\docType{data}
\name{sleep}
\alias{sleep}
\title{Mammal Sleep}
\format{
A data frame with 62 observations (rows) and 11 variables (columns). Each row is a mammal species.
\describe{
\item{species}{A numeric column with the ...}
\item{body_wt}{A numeric column with the ...}
\item{brain_wt}{A numeric column with the ...}
\item{nondream_sleep}{A numeric column with the ...}
\item{dream_sleep}{A numeric column with the ...}
\item{total_sleep}{A numeric column with the ...}
\item{lifespan}{A numeric column with the ...}
\item{gestation}{A numeric column with the ...}
\item{predation_idx}{A numeric column with the ...}
\item{sleep_idx}{A numeric column with the ...}
\item{danger_idx}{A numeric column with the ...}
}
}
\source{
The data are from a study by Allison & Cicchetti (1976).
This data and column descriptions were pulled from StatLib under the name (\href{http://lib.stat.cmu.edu/datasets/sleep}{\code{sleep}}).

Allison, T., & Cicchetti, D. V. (1976). Sleep in mammals: ecological and constitutional correlates. Science, 194 (4266), 732-734.
}
\usage{
sleep
}
\description{
description
}
\keyword{datasets}
